refactor(survey,benchmark): extract survey + benchmark modules into dojo/<module>/ [Phase 1,3,4,5]
