+++ dojo/benchmark/__init__.py
@@ -1,0 +1,1 @@
+import dojo.benchmark.admin  # noqa: F401
