+++ dojo/survey/__init__.py
@@ -1,0 +1,1 @@
+import dojo.survey.admin  # noqa: F401
