+++ dojo/survey/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from dojo.survey.models import General_Survey
+
+admin.site.register(General_Survey)
