+++ dojo/survey/models.py
@@ -1,0 +1,181 @@
+import warnings
+from datetime import timedelta
+
+from django.db import models
+from django.utils import timezone
+from django.utils.translation import gettext as _
+from django_extensions.db.models import TimeStampedModel
+from polymorphic.base import ManagerInheritanceWarning
+from polymorphic.managers import PolymorphicManager
+from polymorphic.models import PolymorphicModel
+
+with warnings.catch_warnings(action="ignore", category=ManagerInheritanceWarning):
+    class Question(PolymorphicModel, TimeStampedModel):
+
+        """Represents a question."""
+
+        class Meta:
+            ordering = ["order"]
+
+        order = models.PositiveIntegerField(default=1,
+                                            help_text=_("The render order"))
+
+        optional = models.BooleanField(
+            default=False,
+            help_text=_("If selected, user doesn't have to answer this question"))
+
+        text = models.TextField(blank=False, help_text=_("The question text"), default="")
+        objects = models.Manager()
+        polymorphic = PolymorphicManager()
+
+        def __str__(self):
+            return self.text
+
+
+class TextQuestion(Question):
+
+    """Question with a text answer"""
+
+    objects = PolymorphicManager()
+
+    def get_form(self):
+        """Returns the form for this model"""
+        from dojo.survey.ui.forms import TextQuestionForm  # noqa: PLC0415 -- lazy import, avoids circular dependency
+        return TextQuestionForm
+
+
+class Choice(TimeStampedModel):
+
+    """Model to store the choices for multi choice questions"""
+
+    order = models.PositiveIntegerField(default=1)
+
+    label = models.TextField(default="")
+
+    class Meta:
+        ordering = ["order"]
+
+    def __str__(self):
+        return self.label
+
+
+class ChoiceQuestion(Question):
+
+    """
+    Question with answers that are chosen from a list of choices defined
+    by the user.
+    """
+
+    multichoice = models.BooleanField(default=False,
+                                      help_text=_("Select one or more"))
+    choices = models.ManyToManyField("dojo.Choice")
+    objects = PolymorphicManager()
+
+    def get_form(self):
+        """Returns the form for this model"""
+        from dojo.survey.ui.forms import ChoiceQuestionForm  # noqa: PLC0415 -- lazy import, avoids circular dependency
+        return ChoiceQuestionForm
+
+
+# meant to be a abstract survey, identified by name for purpose
+class Engagement_Survey(models.Model):
+    name = models.CharField(max_length=200, null=False, blank=False,
+                            editable=True, default="")
+    description = models.TextField(editable=True, default="")
+    questions = models.ManyToManyField("dojo.Question")
+    active = models.BooleanField(default=True)
+
+    class Meta:
+        verbose_name = _("Engagement Survey")
+        verbose_name_plural = "Engagement Surveys"
+        ordering = ("-active", "name")
+
+    def __str__(self):
+        return self.name
+
+
+# meant to be an answered survey tied to an engagement
+
+class Answered_Survey(models.Model):
+    # tie this to a specific engagement
+    engagement = models.ForeignKey("dojo.Engagement", related_name="engagement+",
+                                   null=True, blank=False, editable=True,
+                                   on_delete=models.CASCADE)
+    # what surveys have been answered
+    survey = models.ForeignKey("dojo.Engagement_Survey", on_delete=models.CASCADE)
+    assignee = models.ForeignKey("dojo.Dojo_User", related_name="assignee",
+                                  null=True, blank=True, editable=True,
+                                  default=None, on_delete=models.RESTRICT)
+    # who answered it
+    responder = models.ForeignKey("dojo.Dojo_User", related_name="responder",
+                                  null=True, blank=True, editable=True,
+                                  default=None, on_delete=models.RESTRICT)
+    completed = models.BooleanField(default=False)
+    answered_on = models.DateField(null=True)
+
+    class Meta:
+        verbose_name = _("Answered Engagement Survey")
+        verbose_name_plural = _("Answered Engagement Surveys")
+
+    def __str__(self):
+        return self.survey.name
+
+
+def default_expiration():
+    return timezone.now() + timedelta(days=7)
+
+
+class General_Survey(models.Model):
+    survey = models.ForeignKey("dojo.Engagement_Survey", on_delete=models.CASCADE)
+    num_responses = models.IntegerField(default=0)
+    generated = models.DateTimeField(auto_now_add=True, null=True)
+    expiration = models.DateTimeField(default=default_expiration)
+
+    class Meta:
+        verbose_name = _("General Engagement Survey")
+        verbose_name_plural = _("General Engagement Surveys")
+
+    def __str__(self):
+        return self.survey.name
+
+    def clean(self):
+        if self.expiration and timezone.is_naive(self.expiration):
+            self.expiration = timezone.make_aware(self.expiration)
+
+
+with warnings.catch_warnings(action="ignore", category=ManagerInheritanceWarning):
+    class Answer(PolymorphicModel, TimeStampedModel):
+
+        """Base Answer model"""
+
+        question = models.ForeignKey("dojo.Question", on_delete=models.CASCADE)
+
+        answered_survey = models.ForeignKey("dojo.Answered_Survey",
+                                            null=False,
+                                            blank=False,
+                                            on_delete=models.CASCADE)
+        objects = models.Manager()
+        polymorphic = PolymorphicManager()
+
+
+class TextAnswer(Answer):
+    answer = models.TextField(
+        blank=False,
+        help_text=_("The answer text"),
+        default="")
+    objects = PolymorphicManager()
+
+    def __str__(self):
+        return self.answer
+
+
+class ChoiceAnswer(Answer):
+    answer = models.ManyToManyField(
+        "dojo.Choice",
+        help_text=_("The selected choices as the answer"))
+    objects = PolymorphicManager()
+
+    def __str__(self):
+        if len(self.answer.all()):
+            return str(self.answer.all()[0])
+        return "No Response"
